add controller for members
